Move config for consistency
