+++ cmd/metricswatcher/metrics.go
@@ -7,6 +7,8 @@ import (
 	"os"
 	"time"
 
+	"github.com/mattermost/mattermost-load-test-ng/cmd/metricswatcher/config"
+
 	"github.com/mattermost/mattermost-load-test-ng/cmd/metricswatcher/prometheushelper"
 
 	"github.com/mattermost/mattermost-server/v5/mlog"
@@ -19,7 +21,7 @@ type PrometheusQuery struct {
 	Alert       bool
 }
 
-func checkMetrics(configuration *MetricsCheckConfig, queryFile string) {
+func checkMetrics(configuration *config.MetricsCheckConfig, queryFile string) {
 	prometheusQueries := readPrometheusQueriesFile(queryFile)
 	prometheusHelper := createPrometheusHelper(configuration.PrometheusConfiguration.PrometheusURL)
